refactor: q1 tidy up
